refactor(dashboards): Move structs into implementing files
